Extract discoverPeers
