+++ packages/contracts/source/contracts/p2p/peer-discoverer.ts
@@ -1,3 +1,4 @@
 export interface PeerDiscoverer {
+	discoverPeers(pingAll?: boolean): Promise<boolean>;
 	populateSeedPeers(): Promise<void>;
 }
