+++ packages/p2p/source/peer-discoverer.ts
@@ -1,10 +1,14 @@
-import { inject } from "@mainsail/container";
+import { inject, tagged } from "@mainsail/container";
 import { Contracts, Identifiers } from "@mainsail/contracts";
-import { Services, Utils } from "@mainsail/kernel";
+import { Providers, Services, Utils } from "@mainsail/kernel";
 
 export class PeerDiscoverer implements Contracts.P2P.PeerDiscoverer {
 	@inject(Identifiers.Application)
 	private readonly app!: Contracts.Kernel.Application;
+
+	@inject(Identifiers.PluginConfiguration)
+	@tagged("plugin", "p2p")
+	private readonly configuration!: Providers.PluginConfiguration;
 
 	@inject(Identifiers.PeerFactory)
 	private readonly peerFactory!: Contracts.P2P.PeerFactory;
@@ -12,8 +16,66 @@ export class PeerDiscoverer implements Contracts.P2P.PeerDiscoverer {
 	@inject(Identifiers.PeerRepository)
 	private readonly repository!: Contracts.P2P.PeerRepository;
 
+	@inject(Identifiers.PeerCommunicator)
+	private readonly communicator!: Contracts.P2P.PeerCommunicator;
+
 	@inject(Identifiers.LogService)
 	private readonly logger!: Contracts.Kernel.Logger;
+
+	public async discoverPeers(pingAll?: boolean): Promise<boolean> {
+		const maxPeersPerPeer = 50;
+		const ownPeers: Contracts.P2P.Peer[] = this.repository.getPeers();
+		const theirPeers: Contracts.P2P.Peer[] = Object.values(
+			(
+				await Promise.all(
+					Utils.shuffle(this.repository.getPeers())
+						.slice(0, 8)
+						.map(async (peer: Contracts.P2P.Peer) => {
+							try {
+								const hisPeers = await this.communicator.getPeers(peer);
+								return hisPeers || [];
+							} catch (error) {
+								this.logger.debug(`Failed to get peers from ${peer.ip}: ${error.message}`);
+								return [];
+							}
+						}),
+				)
+			)
+				.map((peers) =>
+					Object.fromEntries(
+						Utils.shuffle(peers)
+							.slice(0, maxPeersPerPeer)
+							.map((current: Contracts.P2P.PeerBroadcast) => [current.ip, this.peerFactory(current.ip)]),
+					),
+				)
+				.reduce(
+					(accumulator: object, current: { [ip: string]: Contracts.P2P.Peer }) => ({
+						...accumulator,
+						...current,
+					}),
+					{},
+				),
+		);
+
+		if (pingAll || !this.#hasMinimumPeers() || ownPeers.length < theirPeers.length * 0.75) {
+			await Promise.all(
+				theirPeers.map((p) =>
+					this.app
+						.get<Services.Triggers.Triggers>(Identifiers.TriggerService)
+						.call("validateAndAcceptPeer", { ip: p.ip, options: { lessVerbose: true } }),
+				),
+			);
+			// eslint-disable-next-line @typescript-eslint/no-floating-promises
+			this.#pingPeerPorts(pingAll);
+
+			return true;
+		}
+
+		// eslint-disable-next-line @typescript-eslint/no-floating-promises
+		this.#pingPeerPorts();
+
+		return false;
+	}
 
 	async populateSeedPeers(): Promise<any> {
 		const peerList: Contracts.P2P.PeerData[] = this.app.config("peers").list;
@@ -71,5 +133,29 @@ export class PeerDiscoverer implements Contracts.P2P.PeerDiscoverer {
 		}
 
 		return [];
+	}
+
+	async #pingPeerPorts(pingAll?: boolean): Promise<void> {
+		let peers = this.repository.getPeers();
+		if (!pingAll) {
+			peers = Utils.shuffle(peers).slice(0, Math.floor(peers.length / 2));
+		}
+
+		this.logger.debug(`Checking ports of ${Utils.pluralize("peer", peers.length, true)}.`);
+
+		await Promise.all(peers.map((peer) => this.communicator.pingPorts(peer)));
+	}
+
+	#hasMinimumPeers(): boolean {
+		if (this.configuration.getOptional<boolean>("ignoreMinimumNetworkReach", false)) {
+			this.logger.warning("Ignored the minimum network reach because the relay is in seed mode.");
+
+			return true;
+		}
+
+		return (
+			Object.keys(this.repository.getPeers()).length >=
+			this.configuration.getRequired<number>("minimumNetworkReach")
+		);
 	}
 }
